ffmpeg/qsv: move JPEG encoder test classes to lib
